More type changes, avoiding implicit conversion warnings.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_softfloat@1690 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/ditool/ditool.c
@@ -566,7 +566,7 @@ static void process_inodes_recr(struct ufs_t* ufs, struct i2p_t** inode2path, st
                     if (ft && vfs_access(&dirEntPath, F_OK) != 0) {
                         struct file_t* file = file_open(&dirEntPath, "wb");
                         if (file_is_open(file)) {
-                            size_t size = ntohl(inode.ic_size);
+                            uint32_t size = ntohl(inode.ic_size);
                             uint8_t* buffer = (uint8_t*)malloc(size);
                             ufs_readFile(ufs, &inode, 0, size, buffer);
                             if (file_write(file, 0, buffer, size) != size) {
