+++ src/slirp/rpc/dns.c
@@ -156,7 +156,7 @@ static size_t domain_name(uint8_t* dst, const char* src) {
 }
 
 static char* alloc_ip_addr_str(uint32_t addr, const char* suffix) {
-    int len = 16 + strlen(suffix);
+    size_t len = 16 + strlen(suffix);
     char* result = (char*)malloc(len);
     snprintf(result, len, "%d.%d.%d.%d%s", (addr>>24)&0xFF, (addr>>16)&0xFF, (addr>>8)&0xFF, addr&0xFF, suffix);
     return result;
